``` feat(kernel): 优化内核内存布局和链接脚本配置
- 调整堆大小从16M减至4M，并添加堆位图段管理
- 减少栈大小从64K至32K，页表堆大小从4M至2M
- 新增init_data和init_stack段用于初始化数据和栈
- 添加kspace_uppdpt段管理内核页表
- 添加heap_bitmap段用于堆内存位图管理
- 优化段加载地址对齐和内存布局

fix(interrupts): 修复本地处理器管理器初始化问题

- 移除gPhyPgsMemMgr和gKspacePgsMemMgr的全局前缀
- 使用phymemspace_mgr和KspaceMapMgr直接访问内存管理器

feat(acpi): 增强ACPI管理器功能并修复内存访问问题

- 添加必要的头文件依赖
- 修改Init方法返回值为int类型并添加错误处理
- 修复ACPI表查找逻辑，添加空指针检查
- 重新映射ACPI段到虚拟地址空间
- 添加偏移量管理以正确访问ACPI表

fix(videodriver): 修复安全数字打印功能

- 将数字类型从uint32_t扩展为uint64_t以支持更大数值

feat(efi): 改进EFI运行时服务初始化和内存管理

- 重构EFI运行时服务类的初始化逻辑
- 添加对EFI内存描述符表的正确处理
- 实现运行时服务内存的虚拟地址映射
- 添加错误处理和状态返回值
```

+++ src/drivers/VideoDriver.cpp
@@ -1119,7 +1119,7 @@ int kpnumSecure(void* numptr, int format, int len)
 { 
     if (!numptr) return -1;
     
-    uint32_t num = *(uint32_t*)numptr;
+    uint64_t num = *(uint64_t*)numptr;
     char buffer[64] = {0};
     
     switch (format) {
